fix dp điểm số, add dp field

+++ DoAnTotNghiep/Models/KQHT.cs
@@ -10,7 +10,7 @@
         public string ModulesName { get; set; }
         public string ClassName { get; set; }
         public string ClassCode { get; set; }
-        public int? Score { get; set; }
+        public decimal? Score { get; set; }
         public int ScoreType { get; set; }
         public decimal Credits { get; set; }
